+++ src/app/login/login.component.ts
@@ -3,6 +3,7 @@ import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { LoginService } from '../services/login.service';
 import { Router } from '@angular/router';
 import { DataService } from '../services/data.service';
+import { saveToken } from '../services/helperFunctions';
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
@@ -35,7 +36,7 @@ export class LoginComponent implements OnInit {
         } else {
           this.router.navigate(['']);
           this.login.getLogged(true);
-          this.dataService.saveToken(<string>res);
+          saveToken(<string>res);
           window.alert("Pomyślnie zalogowano");
         }
       }, error => {
