+++ src/app/services/login.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { BehaviorSubject } from 'rxjs';
-import { DataService } from './data.service';
+import { deleteToken, getToken } from './helperFunctions';
 
 @Injectable({
   providedIn: 'root'
@@ -11,19 +11,19 @@ export class LoginService {
   public logged = new BehaviorSubject(false);
   public currentlogged = this.logged.asObservable();
 
-  constructor(private http: HttpClient, private dataService:DataService) { }
+  constructor(private http: HttpClient,) { }
 
   public loginUser(email, pass) {
     return this.http.post('http://localhost/assets/Login', { 'email': email, 'pass': pass });
   }
 
   public checkAuth(token:string | boolean) {
-    let param = new HttpParams().set('token',this.dataService.getToken())
+    let param = new HttpParams().set('token',getToken())
     return this.http.get('http://localhost/assets/CheckAuth', { params:param });
   }
 
   public logoutUser() {
-    this.dataService.deleteToken();
+    deleteToken();
   }
 
   public getLogged(logged) {
