+++ src/app/services/plan.service.ts
@@ -3,6 +3,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { DataService } from './data.service';
 import { switchMap } from 'rxjs/operators';
 import { LoginService } from './login.service';
+import { getToken } from './helperFunctions';
 
 @Injectable({
   providedIn: 'root'
@@ -11,24 +12,24 @@ export class PlanService {
     constructor(private http: HttpClient, private dataService:DataService, private loginService:LoginService){};
 
     public postPlan(genre, value, fat, cuisine, sex) {
-      return this.loginService.checkAuth(this.dataService.getToken()).pipe(
-        switchMap(()=>this.http.post('http://localhost/assets/Plan', { 'genre': genre, 'value': value, 'fat': fat, 'cuisine': cuisine, 'sex': sex, 'token':this.dataService.getToken() }))
+      return this.loginService.checkAuth(getToken()).pipe(
+        switchMap(()=>this.http.post('http://localhost/assets/Plan', { 'genre': genre, 'value': value, 'fat': fat, 'cuisine': cuisine, 'sex': sex, 'token':getToken()}))
       );
       }
     
     public getPlan() {
-      return this.loginService.checkAuth(this.dataService.getToken()).pipe(
+      return this.loginService.checkAuth(getToken()).pipe(
         switchMap(()=>{
-          let param = new HttpParams().set('token',this.dataService.getToken())
+          let param = new HttpParams().set('token',getToken())
           return this.http.get('http://localhost/assets/Plan', {params:param})
         })
       );
     }
 
     public deletePlan(id_plan) {
-      return this.loginService.checkAuth(this.dataService.getToken()).pipe(
+      return this.loginService.checkAuth(getToken()).pipe(
         switchMap(()=>{
-          let param = new HttpParams().set('token',this.dataService.getToken())
+          let param = new HttpParams().set('token',getToken())
           return this.http.delete('http://localhost/assets/Delete_plan', { params:param });
         })
       );
